Added Screenshot
Added Screenshot of the default HelloWorld component Vue App being served on the load balancer for the front end container service.
Overall was a really fun project. If I was able to spend a little more time, I would have liked to write a bash script to deploy new versions of the app container to the ECR repo so that a canary deployment script could be implemented to route a portion of the traffic to a new version of the container and monitor its performance and dynamically adjust the traffic depending on the results. 

Thank you for taking the time to read my code.